backend/requirements.txt
flask==3.0.3
gunicorn==22.0.0
mysql-connector-python==8.0.33